+++ backend/src/main/java/project/backend/global/security/interceptor/WebSocketChannelInterceptor.java
@@ -1,4 +1,4 @@
-package project.backend.global.webconfig;
+package project.backend.global.security.interceptor;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
