+++ backend/src/main/java/project/backend/global/webconfig/WebSocketConfig.java
@@ -14,6 +14,8 @@ import org.springframework.web.socket.config.annotation.StompEndpointRegistry;
 import org.springframework.web.socket.config.annotation.WebSocketMessageBrokerConfigurer;
 import org.springframework.web.socket.config.annotation.WebSocketTransportRegistration;
 import org.springframework.web.socket.handler.WebSocketHandlerDecorator;
+import project.backend.global.security.interceptor.WebSocketChannelInterceptor;
+import project.backend.global.security.interceptor.WebSocketHandShakeInterceptor;
 
 @Configuration
 @EnableWebSocketMessageBroker
