+++ backend/src/main/resources/application.yml
@@ -43,10 +43,10 @@ spring:
   jpa:
     hibernate:
       ddl-auto: create
-#    properties:
-#      hibernate:
-#        show_sql: true
-#        format_sql: true
+    properties:
+      hibernate:
+        show_sql: true
+        format_sql: true
 
 file:
   images:
